+++ example/es/productionorder.go
@@ -31,7 +31,7 @@ func (order *ProductionOrder) Create(name string) error {
 		BagsToProduce: order.BagsToProduce,
 	}
 
-	em := eventsourcing.NewEventMessage(order.AggregateID(), &created, eventsourcing.Uint64(uint64(order.CurrentVersion())))
+	em := eventsourcing.NewEventMessage(order.AggregateID(), &created, eventsourcing.Int64(order.CurrentVersion()))
 	order.Apply(em, true)
 
 	return nil
